refactor: shared fns
