+++ packages/shared/src/assign.ts
@@ -1,0 +1,11 @@
+export function assign(target: any, ...sources: any[]) {
+  for (const source of sources) {
+    for (const key in source) {
+      if (!target?.hasOwnProperty?.(key)) {
+        target[key] = source[key]
+      }
+    }
+  }
+
+  return target
+}
