+++ packages/shared/src/index.ts
@@ -1,5 +1,6 @@
 export * from './arbitrary-value'
 export * from './assert'
+export * from './assign'
 export * from './astish'
 export * from './calc'
 export * from './camelcase-property'
@@ -7,8 +8,6 @@ export * from './capitalize'
 export * from './classname'
 export * from './compact'
 export * from './condition'
-export * from './important'
-export * from './unit-conversion'
 export * from './css-var'
 export * from './deep-set'
 export * from './entries'
@@ -18,12 +17,15 @@ export * from './flatten'
 export * from './get-or-create-set'
 export * from './hash'
 export * from './hypenate-property'
+export * from './important'
 export * from './is-css-function'
 export * from './is-css-unit'
 export * from './is-css-var'
 export * from './memo'
 export * from './merge-props'
+export * from './merge-with'
 export * from './normalize-style-object'
+export * from './omit'
 export * from './pattern-fns'
 export * from './regex'
 export * from './serialize'
@@ -35,4 +37,5 @@ export * from './to-json'
 export * from './traverse'
 export * from './typegen'
 export * from './uniq'
+export * from './unit-conversion'
 export * from './walk-object'
